refactor: moving functions into separate files for better code structure
