+++ app/src/main/java/com/lab5/wisebites/Profile.kt
@@ -5,16 +5,30 @@ import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.lab5.wisebites.databinding.ActivityProfileBinding
+import com.lab5.wisebites.utils.BottomNavigationHandler
 
 class Profile : AppCompatActivity() {
+
+    private lateinit var binding: ActivityProfileBinding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_profile)
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
-            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
-            insets
-        }
+
+        binding = ActivityProfileBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        // Set the default selected item in Navigation Menu
+        binding.bnMenu.selectedItemId = R.id.i_profile
+
+        // Items Selection Handler
+        BottomNavigationHandler.handleNavigation(this, binding.bnMenu)
+    }
+
+    override fun onRestart() {
+        super.onRestart()
+        // Set the default selected item in Navigation Menu
+        binding.bnMenu.selectedItemId = R.id.i_profile
     }
 }
