+++ app/src/main/java/com/lab5/wisebites/SearchActivity.kt
@@ -1,10 +1,10 @@
 package com.lab5.wisebites
 
-import android.content.Intent
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import com.lab5.wisebites.databinding.ActivitySearchBinding
+import com.lab5.wisebites.utils.BottomNavigationHandler
 
 class SearchActivity : AppCompatActivity() {
 
@@ -21,33 +21,7 @@ class SearchActivity : AppCompatActivity() {
         binding.bnMenu.selectedItemId = R.id.i_search
 
         // Items Selection Handler
-        binding.bnMenu.setOnItemSelectedListener { item ->
-            when (item.itemId) {
-                R.id.i_home -> {
-                    intent = Intent(this, HomeActivity::class.java)
-                    intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_NEW_TASK)
-                    startActivity(intent)
-                    true
-                }
-                R.id.i_search -> {
-                    intent = Intent(this, SearchActivity::class.java)
-                    intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_NEW_TASK)
-                    startActivity(intent)
-                    true
-                }
-                R.id.i_bookmark -> {
-                    intent = Intent(this, BookmarkActivity::class.java)
-                    intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_NEW_TASK)
-                    startActivity(intent)
-                    true
-                }
-                R.id.i_profile -> {
-                    // startActivity(Intent(this, ProfileActivity::class.java))
-                    true
-                }
-                else -> false
-            }
-        }
+        BottomNavigationHandler.handleNavigation(this, binding.bnMenu)
     }
 
     override fun onRestart() {
