+++ app/src/main/res/layout/activity_profile.xml
@@ -82,4 +82,15 @@
         app:cornerRadius="10dp"
         android:text="@string/btn_logOut"/>
 
+    <com.google.android.material.bottomnavigation.BottomNavigationView
+        android:id="@+id/bn_menu"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        app:menu="@menu/bottom_nav_menu"
+        android:background="@drawable/bottom_nav_bg"
+        app:labelVisibilityMode="unlabeled"
+        app:itemIconTint="@color/nav_item_color_selector"
+        app:itemActiveIndicatorStyle="@color/white"
+        app:layout_constraintBottom_toBottomOf="parent"/>
+
     </androidx.constraintlayout.widget.ConstraintLayout>
