+++ app/src/main/res/layout/item_recipes.xml
@@ -79,7 +79,7 @@
                 android:id="@+id/iv_recipe_bookmark"
                 android:layout_width="15dp"
                 android:layout_height="19dp"
-                android:src="@drawable/ic_bookmark_2"
+                android:src="@drawable/recipe_bookmark_selector"
                 android:backgroundTint="@color/white"
                 android:layout_marginEnd="11dp"
                 app:layout_constraintEnd_toEndOf="parent"
